feat: Separate out OsPlacesDpaDto class
